svc-ai-interview/requirements.txt
# Core FastAPI and async support
fastapi==0.115.9
uvicorn[standard]==0.34.3
starlette==0.45.3

# LangChain for AI processing
langchain==0.3.25
langchain-core==0.3.63
langchain-openai==0.3.18
langchain-text-splitters==0.3.8

# AI provider integrations
openai==1.82.1
anthropic>=0.40
google-generativeai>=0.8
langchain-anthropic>=0.3
langchain-google-genai>=2.0

# Configuration and Environment
pydantic==2.11.5
python-dotenv==1.1.0

# HTTP Client for file downloads
requests==2.32.5
httpx==0.28.1

# File processing and parsing
PyMuPDF==1.26.0
python-docx==1.1.2

# Audio processing (for interview transcripts)
openai-whisper==20240930

# Standard library extensions
typing-extensions==4.13.2
